cxx-qt-gen: split rust/types and cpp/types into a cxx utils module
This allows us to have a separate module for independent helpers
for dealing with syn, CXX, and C++ together.

Related to #404
