+++ crates/cxx-qt-gen/src/generator/mod.rs
@@ -6,3 +6,4 @@
 pub mod cpp;
 pub mod naming;
 pub mod rust;
+pub mod utils;
