+++ crates/cxx-qt-gen/src/generator/utils/mod.rs
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: 2023 Klarälvdalens Datakonsult AB, a KDAB Group company <info@kdab.com>
+// SPDX-FileContributor: Andrew Hayzen <andrew.hayzen@kdab.com>
+//
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! This module provides utils for working with syn, CXX, and C++.
+//!
+//! Such as converting a [syn::Type] into a C++ string, or determining
+//! if a type is unsafe in a CXX bridge.
+//!
+//! The idea of this module is that it should be independent as
+//! these methods could be split out into a cxx-utils crate later on.
+
+pub(crate) mod cpp;
+pub(crate) mod rust;
